+++ engine/source/openborscript/binding.c
@@ -225,7 +225,7 @@ HRESULT openbor_set_binding_property(ScriptVariant **varlist, ScriptVariant **pr
 
     // Value carriers to apply on properties after
     // taken from argument.
-    int         temp_int;
+    LONG         temp_int;
 
     // Map string property name to a
     // matching integer constant.
